# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RLS Automation Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Stats {
  totalCalls: number;
  transfers: number;
  rickTransfers: number;
  vedderTransfers: number;
  noAnswers: number;
  voicemails: number;
  hangups: number;
  nextTransferTo: string;
}

export interface MakeCallRequest {
  phone: string;
  businessName: string;
}

export interface CallResponse {
  success?: boolean;
  callSid?: string;
  error?: string;
}

export type QueueItemStatus =
  (typeof QueueItemStatus)[keyof typeof QueueItemStatus];

export const QueueItemStatus = {
  pending: "pending",
  calling: "calling",
  completed: "completed",
  failed: "failed",
} as const;

export interface QueueItem {
  id: string;
  name: string;
  phone: string;
  status: QueueItemStatus;
  callSid?: string;
  startedAt?: string;
  error?: string;
}

export interface Queue {
  items: QueueItem[];
  active: boolean;
  currentIndex: number;
  paused: boolean;
  completed?: number;
  failed?: number;
}

export type AddToQueueRequestLeadsItem = {
  name: string;
  phone: string;
};

export interface AddToQueueRequest {
  leads: AddToQueueRequestLeadsItem[];
}

export interface AddToQueueResponse {
  success?: boolean;
  added?: number;
  error?: string;
}

export interface ActionResponse {
  success?: boolean;
  error?: string;
}

export interface TranscriptLine {
  speaker: string;
  text: string;
  time?: string;
}

export interface Transcript {
  id: string;
  businessName: string;
  outcome: string;
  transferredTo?: string;
  transcript?: TranscriptLine[];
  duration?: number;
  createdAt: string;
}

export interface Learnings {
  insights: string[];
  lastUpdated?: string | null;
  totalAnalyzed: number;
}

export interface TrainingVideo {
  videoId: string;
  title: string;
  channel: string;
  url?: string;
  addedAt?: string;
  techniques?: string[];
}

export interface Training {
  videos: TrainingVideo[];
  techniques: string[];
  lastUpdated?: string | null;
}

export interface AddTrainingRequest {
  url: string;
}

export interface AddTrainingResponse {
  success?: boolean;
  title?: string;
  techniques?: string[];
  error?: string;
}

export interface SarahStatus {
  sarahEnabled: boolean;
  sarahPausedAt?: string | null;
  enabled?: boolean;
}

export type LeadsResponseLeadsItem = {
  name?: string;
  phone?: string;
  notes?: string;
  contactedBy?: string;
  status?: string;
};

export interface LeadsResponse {
  leads: LeadsResponseLeadsItem[];
}

export interface Callback {
  id: string;
  businessName: string;
  phone: string;
  requestedTime: string;
  repName: string;
  createdAt?: string;
  status?: string;
}