# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RLS Automation Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /call:
    post:
      operationId: makeCall
      tags: [calls]
      summary: Initiate a call to a business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MakeCallRequest"
      responses:
        "200":
          description: Call initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallResponse"
...
  /queue/add:
    post:
      operationId: addToQueue
      tags: [queue]
      summary: Add leads to call queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToQueueRequest"
      responses:
        "200":
          description: Leads added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddToQueueResponse"
...
  /training/add:
    post:
      operationId: addTraining
      tags: [training]
      summary: Add a YouTube training video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTrainingRequest"
      responses:
        "200":
          description: Training added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddTrainingResponse"
...
components:
  schemas:
...
    MakeCallRequest:
      type: object
      properties:
        phone:
          type: string
        businessName:
          type: string
      required:
        - phone
        - businessName
    CallResponse:
      type: object
      properties:
        success:
          type: boolean
        callSid:
          type: string
        error:
          type: string
...
    AddToQueueRequest:
      type: object
      properties:
        leads:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              phone:
                type: string
            required:
              - name
              - phone
      required:
        - leads
    AddToQueueResponse:
      type: object
      properties:
        success:
          type: boolean
        added:
          type: integer
        error:
          type: string
...
    AddTrainingRequest:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    AddTrainingResponse:
      type: object
      properties:
        success:
          type: boolean
        title:
          type: string
        techniques:
          type: array
          items:
            type: string
        error:
          type: string
...