# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RLS Automation Hub API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get call statistics
 */
export const GetStatsResponse = zod.object({
  totalCalls: zod.number(),
  transfers: zod.number(),
  rickTransfers: zod.number(),
  vedderTransfers: zod.number(),
  noAnswers: zod.number(),
  voicemails: zod.number(),
  hangups: zod.number(),
  nextTransferTo: zod.string(),
});

/**
 * @summary Initiate a call to a business
 */
export const MakeCallBody = zod.object({
  phone: zod.string(),
  businessName: zod.string(),
});

export const MakeCallResponse = zod.object({
  success: zod.boolean().optional(),
  callSid: zod.string().optional(),
  error: zod.string().optional(),
});

/**
 * @summary Get current queue state
 */
export const GetQueueResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      phone: zod.string(),
      status: zod.enum(["pending", "calling", "completed", "failed"]),
      callSid: zod.string().optional(),
      startedAt: zod.string().optional(),
      error: zod.string().optional(),
    }),
  ),
  active: zod.boolean(),
  currentIndex: zod.number(),
  paused: zod.boolean(),
  completed: zod.number().optional(),
  failed: zod.number().optional(),
});

/**
 * @summary Add leads to call queue
 */
export const AddToQueueBody = zod.object({
  leads: zod.array(
    zod.object({
      name: zod.string(),
      phone: zod.string(),
    }),
  ),
});

export const AddToQueueResponse = zod.object({
  success: zod.boolean().optional(),
  added: zod.number().optional(),
  error: zod.string().optional(),
});

/**
 * @summary Start processing the queue
 */
export const StartQueueResponse = zod.object({
  success: zod.boolean().optional(),
  error: zod.string().optional(),
});

/**
 * @summary Pause queue processing
 */
export const PauseQueueResponse = zod.object({
  success: zod.boolean().optional(),
  error: zod.string().optional(),
});

/**
 * @summary Resume queue processing
 */
export const ResumeQueueResponse = zod.object({
  success: zod.boolean().optional(),
  error: zod.string().optional(),
});

/**
 * @summary Clear the queue
 */
export const ClearQueueResponse = zod.object({
  success: zod.boolean().optional(),
  error: zod.string().optional(),
});

/**
 * @summary Get call transcripts
 */
export const GetTranscriptsResponseItem = zod.object({
  id: zod.string(),
  businessName: zod.string(),
  outcome: zod.string(),
  transferredTo: zod.string().optional(),
  transcript: zod
    .array(
      zod.object({
        speaker: zod.string(),
        text: zod.string(),
        time: zod.string().optional(),
      }),
    )
    .optional(),
  duration: zod.number().optional(),
  createdAt: zod.string(),
});
export const GetTranscriptsResponse = zod.array(GetTranscriptsResponseItem);

/**
 * @summary Get AI learnings from calls
 */
export const GetLearningsResponse = zod.object({
  insights: zod.array(zod.string()),
  lastUpdated: zod.string().nullish(),
  totalAnalyzed: zod.number(),
});

/**
 * @summary Get training videos and techniques
 */
export const GetTrainingResponse = zod.object({
  videos: zod.array(
    zod.object({
      videoId: zod.string(),
      title: zod.string(),
      channel: zod.string(),
      url: zod.string().optional(),
      addedAt: zod.string().optional(),
      techniques: zod.array(zod.string()).optional(),
    }),
  ),
  techniques: zod.array(zod.string()),
  lastUpdated: zod.string().nullish(),
});

/**
 * @summary Add a YouTube training video
 */
export const AddTrainingBody = zod.object({
  url: zod.string(),
});

export const AddTrainingResponse = zod.object({
  success: zod.boolean().optional(),
  title: zod.string().optional(),
  techniques: zod.array(zod.string()).optional(),
  error: zod.string().optional(),
});

/**
 * @summary Get Sarah AI status
 */
export const GetSarahStatusResponse = zod.object({
  sarahEnabled: zod.boolean(),
  sarahPausedAt: zod.string().nullish(),
  enabled: zod.boolean().optional(),
});

/**
 * @summary Enable Sarah AI
 */
export const EnableSarahResponse = zod.object({
  sarahEnabled: zod.boolean(),
  sarahPausedAt: zod.string().nullish(),
  enabled: zod.boolean().optional(),
});

/**
 * @summary Disable Sarah AI
 */
export const DisableSarahResponse = zod.object({
  sarahEnabled: zod.boolean(),
  sarahPausedAt: zod.string().nullish(),
  enabled: zod.boolean().optional(),
});

/**
 * @summary Send test follow-up texts now
 */
export const TestTexterResponse = zod.object({
  success: zod.boolean().optional(),
  error: zod.string().optional(),
});

/**
 * @summary Get today's follow-up leads from CRM
 */
export const GetLeadsResponse = zod.object({
  leads: zod.array(
    zod.object({
      name: zod.string().optional(),
      phone: zod.string().optional(),
      notes: zod.string().optional(),
      contactedBy: zod.string().optional(),
      status: zod.string().optional(),
    }),
  ),
});

/**
 * @summary Get scheduled callbacks
 */
export const GetCallbacksResponseItem = zod.object({
  id: zod.string(),
  businessName: zod.string(),
  phone: zod.string(),
  requestedTime: zod.string(),
  repName: zod.string(),
  createdAt: zod.string().optional(),
  status: zod.string().optional(),
});
export const GetCallbacksResponse = zod.array(GetCallbacksResponseItem);