# lib/api-zod/src/generated/types/addToQueueRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RLS Automation Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { AddToQueueRequestLeadsItem } from "./addToQueueRequestLeadsItem";

export interface AddToQueueRequest {
  leads: AddToQueueRequestLeadsItem[];
}