# lib/api-zod/src/generated/types/callResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RLS Automation Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CallResponse {
  success?: boolean;
  callSid?: string;
  error?: string;
}